requirements.txt
streamlit==1.30.0
tensorflow-cpu==2.14.0
numpy==1.24.3
opencv-python
matplotlib
pillow
scikit-learn==1.3.2
